aircraft_detection_advanced/requirements.txt
# Core ML/CV
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0,<3.0
timm>=0.9.0

# Tracking
supervision>=0.16.0  # includes ByteTrack
norfair>=2.2.0

# Geospatial
shapely>=2.0.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# Utilities
pyyaml>=6.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
Pillow>=10.0.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Logging
loguru>=0.7.0
